Code refactoring, now using multiple files
